move export-related functions back to contrib side
