+++ src/support/base64.h
@@ -293,40 +293,6 @@ class Base64OutStream : public dmlc::Stream {
     }
   }
 };
-
-inline size_t b64strlen(const std::string b64str) {
-  ICHECK(b64str.size() % 4 == 0) << "invalid base64 encoding";
-  size_t length = b64str.size() / 4 * 3;
-  if (b64str[b64str.size() - 2] == '=') {
-    length -= 2;
-  } else if (b64str[b64str.size() - 1] == '=') {
-    length -= 1;
-  }
-  return length;
-}
-
-inline void b64decode(const std::string b64str, uint8_t* ret) {
-  size_t index = 0;
-  const auto length = b64str.size();
-  for (size_t i = 0; i < length; i += 4) {
-    int8_t ch0 = base64::DecodeTable[(int32_t)b64str[i]];
-    int8_t ch1 = base64::DecodeTable[(int32_t)b64str[i + 1]];
-    int8_t ch2 = base64::DecodeTable[(int32_t)b64str[i + 2]];
-    int8_t ch3 = base64::DecodeTable[(int32_t)b64str[i + 3]];
-    uint8_t st1 = (ch0 << 2) + (ch1 >> 4);
-    ret[index++] = st1;
-    if (b64str[i + 2] != '=') {
-      uint8_t st2 = ((ch1 & 0b1111) << 4) + (ch2 >> 2);
-      ret[index++] = st2;
-      if (b64str[i + 3] != '=') {
-        uint8_t st3 = ((ch2 & 0b11) << 6) + ch3;
-        ret[index++] = st3;
-      }
-    }
-  }
-  ICHECK(b64strlen(b64str) == index) << "base64 decoding fails";
-}
-
 }  // namespace support
 }  // namespace tvm
 #endif  // TVM_SUPPORT_BASE64_H_
